Additional move things to the correct place
